use useEffect and useCallBack to trigger the fetch function whenever the component is loaded for the first time and whenever the button is clicked
First, instead of calling the function directly
inside the component, is good to create a useEffect
hook to call the function. Leaving the dependencies
array empty, will trigger the function the first
time the component is loaded.

However, it's a good practice to add everything used
inside the useEffect hook in the dependencies array.
But by doing that, you might create a infinite loop
because every time the component is loaded, the function
pointer changes because a new variable is assigned,
a function in JS is an object

The best practice is to add the function inside an
anonymous functions inside a useCallBack hook.
That way, the function is created and is stored inside a
useCallBack hook, so it won't re-render every time
the component is loaded without a reason